Updated for Java 14 & new hardware
Previous perf work was done on Java 12. Now on Java 14.0.2, and some of the micro opts didn't make a difference anymore.

Tweaked the warmup iterations for the perf tests to ensure the hotspot compiler has time to work.

Updated all perf results for runs on my current laptop, added HW specs to the comments.

Updated a few comments here and there for clarity (hopefully!!)