No-alloc check method
Reduces GC pressure greatly for big runs. The improvement is worth it. This example is here to show students how to play all the games and expose them to Knuth's algorithm, so a little extra complexity in the check method is fine since they should be writing their own anyway.