C++ Mastermind
This is a C++ version of Mastermind, with the same test cases and algorithms to play all games as the Java version. The current scoring function is ported from Java to C++. Additionally, a series of optimizations are provided to the scoring function which results in a by-hand vectorization using SSE that is O(1) time, and extremely fast compared to all of the other previous implementations.