Allow up to 15 colors and 15 pins for Java Mastermind
Cleaned up a smidge to allow up to 15 colors and/or pins.

Ditched the score cache. This was a big win in the old days, but now with faster scoring functions and especially the Strategy changes it was just either noise, or slower. Code is simpler without it, and my goal for the Java example is relative simplicity. For uber speed, see the C++ versions.