C# version of the Levenshtein distance problem for AP CS.
This is a departure from my previous versions in Java, C++, and Python. This one searches for paths from both ends, and is a shorter, simpler implementation to try to show some of the fundmental concepts in this problem.

Credit to Tristen Yim, a student in Miyoshi's AP CS class, for the idea of searching from both sides at the same time.