Updates for clarity and minor tweaks
I haven't revisited this code in ~3 years, and looking at it again recently with a student I realized a little update was in order.

Renamed a bunch of variable to hopefully clarify some of the concepts.

Fixed a bunch of typos.

Only very minor gameplay tweaks: simplified one or two little things, again for clarity.

Updated the results with current numbers.